Refactor + Add TeleOp Sample
