+++ examples/src/main/java/org/firstinspires/ftc/teamcode/PedroCommandSample/PedroTeleOpExample.java
@@ -1,0 +1,31 @@
+package org.firstinspires.ftc.teamcode.PedroCommandSample;
+
+
+import com.pedropathing.follower.Follower;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.seattlesolvers.solverslib.command.CommandOpMode;
+
+@TeleOp
+public class PedroTeleOpExample extends CommandOpMode {
+    Follower follower = new Follower(hardwareMap);
+
+    @Override
+    public void initialize() {
+        super.reset();
+
+        follower.startTeleopDrive();
+    }
+
+    @Override
+    public void run() {
+        super.run();
+
+        /* Robot-Centric Drive
+        follower.setTeleOpMovementVectors(-gamepad1.left_stick_y, -gamepad1.left_stick_x, -gamepad1.right_stick_x, true);
+        */
+
+        // Field-Centric Drive
+        follower.setTeleOpMovementVectors(-gamepad1.left_stick_y, -gamepad1.left_stick_x, -gamepad1.right_stick_x, false);
+        follower.update();
+    }
+}
